backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
sqlglot>=23
openai>=1.40
langchain>=0.2
langchain-openai>=0.1
databricks-sql-connector>=3.1
pytest>=8.0
